Add clear btn to search/filter and menu clear unselected
Added a button to clear the filter options on the main window. Also
updated the menu code so right click option is to remove unselected
items not selected items. And last but not least added code to check the
ingredient list when you click to send an email and if no ingredients
are selected emails the entire list but is there is at least one
selected only includes the selected ingredients in the email.